backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
firebase-admin==6.4.0
sentence-transformers==2.6.1
scikit-learn==1.4.1.post1
numpy==1.26.4
httpx==0.27.0
python-multipart==0.0.9
google-generativeai>=0.8.3
